+++ UI/_builder/Tmpl/core/Resolvers.ts
@@ -5,7 +5,15 @@
  * @file UI/_builder/Tmpl/core/Resolvers.ts
  */
 
-import { isLogicalPath } from 'UI/_builder/Tmpl/core/Path';
+/**
+ * Empty string constant.
+ */
+const EMPTY_STRING = '';
+
+/**
+ * Special prefix for component and object options.
+ */
+const WS_PREFIX_PATTERN = /^ws:/i;
 
 /**
  * Separator for RequireJS plugins.
@@ -28,14 +36,9 @@ const LOGICAL_PATH_SEPARATOR = '.';
 const INTER_PATH_SEPARATOR = ':';
 
 /**
- * Empty string constant.
+ * Pattern for extension of template file path.
  */
-const EMPTY_STRING = '';
-
-/**
- * Special prefix for component and object options.
- */
-const WS_PREFIX_PATTERN = /^ws:/i;
+const TEMPLATE_FILE_EXTENSION_PATTER = /\.(tmpl|wml)/i;
 
 /**
  * Special UI-module names that do not obey standard of naming UI-modules.
@@ -45,25 +48,6 @@ const SPECIAL_UI_MODULE_NAMES = [
    'SBIS3.ENGINE'
 ];
 
-/**
- * Pattern for extension of template file path.
- */
-const TEMPLATE_FILE_EXTENSION_PATTER = /\.(tmpl|wml)/i;
-
-/**
- * Get RequireJS plugin flag by plugin name.
- * @param plugin {string} RequireJS plugin name.
- * @returns {RequireJSPlugins} Returns found RequireJS plugin flag.
- */
-function getPluginFlag(plugin: string): RequireJSPlugins {
-   const pluginName = plugin.toUpperCase();
-   if (RequireJSPlugins[pluginName]) {
-      return RequireJSPlugins[pluginName];
-   }
-   throw new Error(
-      `обнаружен неизвестный плагин "${plugin}" RequireJS`
-   );
-}
 
 /**
  * Find special UI module name if input name starts with its.
@@ -81,20 +65,252 @@ function findSpecialUIModulePrefix(physicalPath: string): string {
 }
 
 /**
- * Extract RequireJS plugins from full path.
- * @param fullPath {string} Full path to source with RequireJS plugins.
+ * @todo
  */
-function extractPlugins(fullPath: string): { plugins: RequireJSPlugins; path: string; } {
-   const parts = fullPath.split(REQUIRE_JS_PLUGIN_SEPARATOR);
-   const path = parts.pop();
-   let plugins = RequireJSPlugins.NONE;
-   for (let index = 0; index < parts.length; ++index) {
-      plugins = plugins | getPluginFlag(parts[index]);
+export interface IPath {
+
+   /**
+    * @todo
+    */
+   getFullPath(): string;
+
+   /**
+    * @todo
+    */
+   getFullPhysicalPath(): string;
+
+   /**
+    * @todo
+    */
+   getLogicalPath(): string[];
+
+   /**
+    * @todo
+    */
+   hasLogicalPath(): boolean;
+
+   /**
+    * @todo
+    */
+   hasPlugins(): boolean;
+}
+
+
+/**
+ * @todo
+ */
+class Path implements IPath {
+
+   /**
+    * @todo
+    */
+   private readonly physicalPath: string[];
+
+   /**
+    * @todo
+    */
+   private readonly logicalPath: string[];
+
+   /**
+    * @todo
+    */
+   private readonly plugins: string[];
+
+   /**
+    * @todo
+    * @param physicalPath {string[]} @todo
+    * @param logicalPath {string[]} @todo
+    * @param plugins {string[]} @todo
+    */
+   constructor(physicalPath: string[], logicalPath: string[], plugins: string[]) {
+      this.physicalPath = physicalPath;
+      this.logicalPath = logicalPath;
+      this.plugins = plugins;
    }
+
+   /**
+    * @todo
+    */
+   getFullPath(): string {
+      const fullPhysicalPath = this.getFullPhysicalPath();
+      const logicalPath = this.logicalPath.length > 0
+         ? INTER_PATH_SEPARATOR + this.logicalPath.join(LOGICAL_PATH_SEPARATOR)
+         : EMPTY_STRING;
+      return fullPhysicalPath + logicalPath;
+   }
+
+   /**
+    * @todo
+    */
+   getFullPhysicalPath(): string {
+      const plugins = this.plugins.length > 0
+         ? this.plugins.join(REQUIRE_JS_PLUGIN_SEPARATOR) + REQUIRE_JS_PLUGIN_SEPARATOR
+         : EMPTY_STRING;
+      const physicalPath = this.physicalPath.join(PHYSICAL_PATH_SEPARATOR);
+      return plugins + physicalPath;
+   }
+
+   /**
+    * @todo
+    */
+   getLogicalPath(): string[] {
+      return this.logicalPath;
+   }
+
+   /**
+    * @todo
+    */
+   hasLogicalPath(): boolean {
+      return this.logicalPath.length > 0;
+   }
+
+   /**
+    * @todo
+    */
+   hasPlugins(): boolean {
+      return this.plugins.length > 0;
+   }
+}
+
+/**
+ * @todo
+ * @param path {string} @todo
+ */
+function splitPlugins(path: string): { plugins: string[]; fullPath: string; } {
+   const plugins = path.split(REQUIRE_JS_PLUGIN_SEPARATOR);
+   const fullPath = plugins.pop();
    return {
       plugins,
-      path
+      fullPath
    };
+}
+
+/**
+ * @todo
+ * @param componentName {string} @todo
+ */
+export function parseComponentName(componentName: string): IPath {
+   // FIXME: Check if tag node name is component name with prefix "ws:"
+   const cleanPath = componentName.replace(WS_PREFIX_PATTERN, EMPTY_STRING);
+   const paths = cleanPath.split(INTER_PATH_SEPARATOR);
+   if (paths.length > 2) {
+      throw new Error(
+         `некорректное имя компонента "${cleanPath}" - ожидалось не более 1 COLON(:)-разделителя`
+      );
+   }
+   // TODO: validate paths
+   const physicalPathString = paths.shift();
+   const logicalPathString = paths.length === 1 ? paths.shift() : EMPTY_STRING;
+   const prefix = findSpecialUIModulePrefix(physicalPathString);
+   const physicalPathOffset = prefix !== EMPTY_STRING? prefix.length + 1 : 0;
+   const physicalPath = physicalPathString
+      .slice(physicalPathOffset)
+      .split(LOGICAL_PATH_SEPARATOR);
+   if (prefix !== EMPTY_STRING) {
+      physicalPath.unshift(prefix);
+   }
+   const logicalPath = logicalPathString !== EMPTY_STRING
+      ? logicalPathString.split(LOGICAL_PATH_SEPARATOR)
+      : [ ];
+   return new Path(physicalPath, logicalPath, []);
+}
+
+/**
+ * @todo
+ * @param functionPath {string} @todo
+ */
+export function parseFunctionPath(functionPath: string): IPath {
+   const { plugins, fullPath } = splitPlugins(functionPath);
+   const paths = fullPath.split(INTER_PATH_SEPARATOR);
+   if (paths.length > 2) {
+      throw new Error(
+         `некорректный путь к функции "${fullPath}" - ожидалось не более 1 COLON(:)-разделителя`
+      );
+   }
+   const physicalPathString = paths.shift();
+   if (TEMPLATE_FILE_EXTENSION_PATTER.test(physicalPathString)) {
+      throw new Error(`некорректный путь к функции "${fullPath}" - указано расширение файла`);
+   }
+   const logicalPathString = paths.length === 1 ? paths.shift() : EMPTY_STRING;
+   const prefix = findSpecialUIModulePrefix(physicalPathString);
+   const physicalPathOffset = prefix !== EMPTY_STRING ? prefix.length + 1 : 0;
+   const physicalPath = physicalPathString
+      .slice(physicalPathOffset)
+      .split(PHYSICAL_PATH_SEPARATOR);
+   if (prefix !== EMPTY_STRING) {
+      physicalPath.unshift(prefix);
+   }
+   const logicalPath = logicalPathString !== EMPTY_STRING
+      ? logicalPathString.split(LOGICAL_PATH_SEPARATOR)
+      : [ ];
+   if (physicalPath.length === 0) {
+      throw new Error(
+         `некорректный путь к функции "${fullPath}" - отсутствует физический путь к модулю, в котором находится запрашиваемая функция`
+      );
+   }
+   if (logicalPath.length === 0) {
+      throw new Error(
+         `некорректный путь к функции "${fullPath}" - отсутствует логический путь к функции, по которому функция должна быть разрешена внутри указанного модуля`
+      );
+   }
+   if (plugins.length !== 0) {
+      throw new Error(
+         `некорректный путь к функции "${fullPath}" - использовать плагины RequireJS запрещено`
+      );
+   }
+   return new Path(physicalPath, logicalPath, plugins);
+}
+
+/**
+ * @todo
+ * @param templatePath {string} @todo
+ */
+export function parseTemplatePath(templatePath: string): IPath {
+   const { plugins, fullPath } = splitPlugins(templatePath);
+   const paths = fullPath.split(INTER_PATH_SEPARATOR);
+   if (paths.length > 2) {
+      throw new Error(
+         `некорректный путь к функции "${fullPath}" - ожидалось не более 1 COLON(:)-разделителя`
+      );
+   }
+   // TODO: validate paths
+   const physicalPathString = paths.shift();
+   const logicalPathString = paths.length > 0 ? paths.shift() : EMPTY_STRING;
+   const prefix = findSpecialUIModulePrefix(physicalPathString);
+   const physicalPathOffset = prefix !== EMPTY_STRING ? prefix.length + 1 : 0;
+   const physicalPath = physicalPathString
+      .slice(physicalPathOffset)
+      .split(PHYSICAL_PATH_SEPARATOR);
+   if (prefix !== EMPTY_STRING) {
+      physicalPath.unshift(prefix);
+   }
+   if (physicalPath.length === 0) {
+      throw new Error(
+         `некорректный путь к файлу "${fullPath}" - отсутствует физический путь к модулю, в котором находится запрашиваемый шаблон`
+      );
+   }
+   const logicalPath = logicalPathString !== EMPTY_STRING
+      ? logicalPathString.split(LOGICAL_PATH_SEPARATOR)
+      : [ ];
+   return new Path(physicalPath, logicalPath, plugins);
+}
+
+/**
+ * Fast check if path represents physical path.
+ * @param path {string} Path.
+ * @returns {boolean} Returns true if path contains physical path separator.
+ */
+export function isPhysicalPath(path: string): boolean {
+   return path.indexOf(PHYSICAL_PATH_SEPARATOR) > -1;
+}
+
+/**
+ * Fast check if path represents logical path.
+ * @param path {string} Path.
+ * @returns {boolean} Returns true if path contains logical path separator.
+ */
+export function isLogicalPath(path: string): boolean {
+   return path.indexOf(LOGICAL_PATH_SEPARATOR) > -1 && path.indexOf(PHYSICAL_PATH_SEPARATOR) === -1;
 }
 
 /**
@@ -128,153 +344,18 @@ export enum RequireJSPlugins {
    XML = 1 << 16
 }
 
-/**
- * Interface for complex paths.
- */
-export interface IComplexPath {
-
-   /**
-    * Mounted RequireJS plugins.
-    */
-   plugins: RequireJSPlugins;
-
-   /**
-    * Physical path to module.
-    */
-   physicalPath: string[];
-
-   /**
-    * Logical path to value.
-    */
-   logicalPath: string[];
-}
-
-/**
- * Parse component name into complex path.
- * @param fullPath {string} Component name.
- * @returns {IComplexPath} Returns complex path description for component.
- * @throws {Error} Throws Error in case of invalid path for component.
- */
-export function parseComponentPath(fullPath: string): IComplexPath {
-   const paths = fullPath.split(INTER_PATH_SEPARATOR);
-   if (paths.length > 2) {
-      throw new Error(
-         `некорректное имя компонента "${fullPath}" - ожидалось не более 1 COLON(:)-разделителя`
-      );
-   }
-   // TODO: validate paths
-   const physicalPathString = paths.shift();
-   const logicalPathString = paths.length === 1 ? paths.shift() : EMPTY_STRING;
-   const prefix = findSpecialUIModulePrefix(physicalPathString);
-   const physicalPathOffset = prefix !== EMPTY_STRING? prefix.length + 1 : 0;
-   const physicalPath = physicalPathString
-      .slice(physicalPathOffset)
-      .split(LOGICAL_PATH_SEPARATOR);
-   if (prefix !== EMPTY_STRING) {
-      physicalPath.unshift(prefix);
-   }
-   const logicalPath = logicalPathString !== EMPTY_STRING
-      ? logicalPathString.split(LOGICAL_PATH_SEPARATOR)
-      : [ ];
-   return {
-      plugins: RequireJSPlugins.NONE,
-      physicalPath,
-      logicalPath
-   };
-}
-
-/**
- * Parse function path.
- * @param fullPath {string} Full path to source with RequireJS plugins.
- * @returns {IComplexPath} Returns complex path description for function.
- * @throws {Error} Throws Error in case of invalid path for function.
- */
-export function parseFunctionPath(fullPath: string): IComplexPath {
-   const { plugins, path } = extractPlugins(fullPath);
-   const paths = path.split(INTER_PATH_SEPARATOR);
-   if (paths.length > 2) {
-      throw new Error(
-         `некорректный путь к функции "${fullPath}" - ожидалось не более 1 COLON(:)-разделителя`
-      );
-   }
-   // TODO: validate paths
-   const physicalPathString = paths.shift();
-   if (TEMPLATE_FILE_EXTENSION_PATTER.test(physicalPathString)) {
-      throw new Error(`некорректный путь к функции "${fullPath}" - указано расширение файла`);
-   }
-   const logicalPathString = paths.length === 1 ? paths.shift() : EMPTY_STRING;
-   const prefix = findSpecialUIModulePrefix(physicalPathString);
-   const physicalPathOffset = prefix !== EMPTY_STRING ? prefix.length + 1 : 0;
-   const physicalPath = physicalPathString
-      .slice(physicalPathOffset)
-      .split(PHYSICAL_PATH_SEPARATOR);
-   if (prefix !== EMPTY_STRING) {
-      physicalPath.unshift(prefix);
-   }
-   const logicalPath = logicalPathString !== EMPTY_STRING
-      ? logicalPathString.split(LOGICAL_PATH_SEPARATOR)
-      : [ ];
-   if (physicalPath.length === 0) {
-      throw new Error(
-         `некорректный путь к функции "${fullPath}" - отсутствует физический путь к модулю, в котором находится запрашиваемая функция`
-      );
-   }
-   if (logicalPath.length === 0) {
-      throw new Error(
-         `некорректный путь к функции "${fullPath}" - отсутствует логический путь к функции, по которому функция должна быть разрешена внутри указанного модуля`
-      );
-   }
-   if (plugins !== RequireJSPlugins.NONE) {
-      throw new Error(
-         `некорректный путь к функции "${fullPath}" - использовать плагины RequireJS запрещено`
-      );
-   }
-   return {
-      plugins,
-      physicalPath,
-      logicalPath
-   };
-}
-
-/**
- * Parse template file path.
- * @param fullPath {string} Full path to source with RequireJS plugins.
- * @returns {IComplexPath} Returns complex path description for template file path.
- * @throws {Error} Throws Error in case of invalid path for template file path.
- */
-export function parseTemplatePath(fullPath: string): IComplexPath {
-   const { plugins, path } = extractPlugins(fullPath);
-   // TODO: validate paths
-   const logicalPath = [ ];
-   if (TEMPLATE_FILE_EXTENSION_PATTER.test(path)) {
-      throw new Error(`некорректный путь к файлу "${fullPath}" - указано расширение файла`);
-   }
-   const prefix = findSpecialUIModulePrefix(path);
-   const physicalPathOffset = prefix !== EMPTY_STRING ? prefix.length + 1 : 0;
-   const physicalPath = path
-      .slice(physicalPathOffset)
-      .split(PHYSICAL_PATH_SEPARATOR);
-   if (prefix !== EMPTY_STRING) {
-      physicalPath.unshift(prefix);
-   }
-   if (physicalPath.length === 0) {
-      throw new Error(
-         `некорректный путь к файлу "${fullPath}" - отсутствует физический путь к модулю, в котором находится запрашиваемый шаблон`
-      );
-   }
-   return {
-      plugins,
-      physicalPath,
-      logicalPath
-   };
-}
 
 /**
  * Check if name is valid option name.
  * @param name {string} Option name.
  */
 export function isOption(name: string): boolean {
-   return WS_PREFIX_PATTERN.test(name);
+   if (!WS_PREFIX_PATTERN.test(name)) {
+      return false;
+   }
+   // FIXME: Check if tag node name is component name with prefix "ws:"
+   const cleanName = name.replace(WS_PREFIX_PATTERN, EMPTY_STRING);
+   return !(isLogicalPath(cleanName) && isCapitalized(cleanName));
 }
 
 /**
@@ -291,5 +372,7 @@ export function resolveOption(name: string): string {
  * @returns {boolean} Returns true if name is valid logical path and first letter is capitalized.
  */
 export function isComponent(name: string): boolean {
-   return isLogicalPath(name) && isCapitalized(name);
+   // FIXME: Check if tag node name is component name with prefix "ws:"
+   const cleanName = name.replace(WS_PREFIX_PATTERN, EMPTY_STRING);
+   return isLogicalPath(cleanName) && isCapitalized(cleanName);
 }
